solve homework 4-2

+++ home work 4-2.py
@@ -5,28 +5,28 @@ if num1.isdigit():
     opr = input("please choose an operation")
     if opr == '1' or opr == '+':
         num2 = input("Please enter the second number")
-        print(float(num1) + float(num2))
+        print(int(num1) + int(num2))
 
     elif opr == '2' or opr == '-':
         num2 = input("Please enter the second number")
 
-        print(float(num1) - float(num2))
+        print(int(num1) - int(num2))
 
     elif opr == '3' or opr == '*':
         num2 = input("Please enter the second number")
-        print(float(num1) * float(num2))
+        print(int(num1) * int(num2))
 
     elif opr == '4' or opr == '/':
         num2 = input("Please enter the second number")
-        print(float(num1) / float(num2))
+        print(int(num1) / int(num2))
 
     elif opr == '5' or opr == '^':
         num2 = input("Please enter the second number")
-        print(float(num1) ^ float(num2))
+        print(int(num1) ^ int(num2))
 
     elif opr == '6' or opr == '%':
         num2 = input("Please enter the second number")
-        print(float(num1) % float(num2))
+        print(int(num1) % int(num2))
 
 else:
     print("Error")
